vocaldoc-analysis/requirements.txt
fastapi
uvicorn
praat-parselmouth
numpy
scipy
python-multipart
matplotlib
pytest
